requirements.txt
flask
jinja2
gunicorn
python-dotenv
pymysql